fix(TileView): not showing all participants.
