+++ react/features/filmstrip/reducer.js
@@ -117,10 +117,14 @@ ReducerRegistry.register(
                 horizontalViewDimensions: action.dimensions
             };
         case SET_REMOTE_PARTICIPANTS: {
-            const { visibleParticipantsStartIndex: startIndex, visibleParticipantsEndIndex: endIndex } = state;
-
             state.remoteParticipants = action.participants;
-            state.visibleRemoteParticipants = new Set(state.remoteParticipants.slice(startIndex, endIndex));
+
+            // TODO: implement this on mobile.
+            if (navigator.product !== 'ReactNative') {
+                const { visibleParticipantsStartIndex: startIndex, visibleParticipantsEndIndex: endIndex } = state;
+
+                state.visibleRemoteParticipants = new Set(state.remoteParticipants.slice(startIndex, endIndex));
+            }
 
             return { ...state };
         }
