+++ react/features/filmstrip/subscriber.native.js
@@ -1,0 +1,3 @@
+// @flow
+
+import './subscriber.any';
